+++ lib/in_app_gallery.dart
@@ -1,2 +1,3 @@
 export 'presentation/in_app_gallery_screen/in_app_gallery_screen.dart';
 export 'utils/in_app_gallery_utils.dart';
+export 'widgets/in_app_gallery_video_player.dart';